requirements.txt
Pillow
numpy
opencv-python
scikit-image
scipy
pandas
tqdm